fix(spawn): default dialogs to floating centered placement
Detect XDG dialogs from modal toplevel state or parent/transient relationships and apply built-in dialog behavior unless overridden by user rules.

Dialogs now default to overlap-policy=all, spawn-placement=center, and cluster-participation=float. Parented dialogs center over their parent window, while unparented dialogs fall back to the viewport center.

Verified with
cargo fmt --check,
cargo check -p halley-wl,
cargo test -p halley-wl compositor::spawn::rules --lib,
and cargo test -p halley-wl tiled_cluster_layout_participation_honors_layout_and_float --lib.
